[tested] merge node plugin into unified plugin interface
