+++ pkg/plugins/docsy.go
@@ -10,7 +10,6 @@ import (
 	"strings"
 
 	"github.com/gardener/docforge/pkg/manifest"
-	"github.com/gardener/docforge/pkg/nodeplugins"
 	"github.com/gardener/docforge/pkg/registry"
 )
 
@@ -43,7 +42,7 @@ func (p *DocsyPlugin) Process(*manifest.Node) error {
 }
 
 // ProcessNew processes a node using the new channel-based method
-func (p *DocsyPlugin) ProcessNew(*manifest.Node) []chan nodeplugins.Status {
+func (p *DocsyPlugin) ProcessNew(*manifest.Node) []chan Status {
 	return nil // Not used
 }
