+++ pkg/plugins/downloader/test/README.md
