+++ pkg/plugins/markdown/document_worker.go
@@ -14,9 +14,9 @@ import (
 
 	"github.com/gardener/docforge/cmd/hugo"
 	"github.com/gardener/docforge/pkg/manifest"
-	"github.com/gardener/docforge/pkg/nodeplugins/markdown/frontmatter"
-	"github.com/gardener/docforge/pkg/nodeplugins/markdown/linkresolver"
-	"github.com/gardener/docforge/pkg/nodeplugins/markdown/parser"
+	"github.com/gardener/docforge/pkg/plugins/markdown/frontmatter"
+	"github.com/gardener/docforge/pkg/plugins/markdown/linkresolver"
+	"github.com/gardener/docforge/pkg/plugins/markdown/parser"
 	"github.com/gardener/docforge/pkg/registry"
 	"github.com/gardener/docforge/pkg/registry/repositoryhost"
 	"github.com/gardener/docforge/pkg/writers"
