+++ pkg/plugins/markdown/linkresolver/linkresolverfakes/fake_interface.go
@@ -8,7 +8,7 @@ import (
 	"sync"
 
 	"github.com/gardener/docforge/pkg/manifest"
-	"github.com/gardener/docforge/pkg/nodeplugins/markdown/linkresolver"
+	"github.com/gardener/docforge/pkg/plugins/markdown/linkresolver"
 )
 
 type FakeInterface struct {
