+++ pkg/plugins/markdown/plugin.go
@@ -1,32 +1,29 @@
-// SPDX-FileCopyrightText: 2025 SAP SE or an SAP affiliate company and Gardener contributors
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package plugins
+package markdown
 
 import (
+	"context"
 	"strings"
 
 	"github.com/gardener/docforge/cmd/hugo"
 	"github.com/gardener/docforge/pkg/manifest"
-	"github.com/gardener/docforge/pkg/nodeplugins"
-	"github.com/gardener/docforge/pkg/nodeplugins/markdown"
+	"github.com/gardener/docforge/pkg/plugins"
+	"github.com/gardener/docforge/pkg/plugins/markdown/linkresolver"
 	"github.com/gardener/docforge/pkg/registry"
 	"github.com/gardener/docforge/pkg/writers"
 )
 
-// MarkdownPlugin handles both manifest transformations and node processing for markdown files
-type MarkdownPlugin struct {
+// Plugin handles both manifest transformations and node processing for markdown files
+type Plugin struct {
 	registry           registry.Interface
 	hugo               hugo.Hugo
 	writer             writers.Writer
 	skipLinkValidation bool
-	markdownProcessor  nodeplugins.Interface // Created lazily in FinalNodeStructure
+	documentWorker     *Worker // Created in FinalNodeStructure
 }
 
-// NewMarkdownPlugin creates a new markdown plugin
-func NewMarkdownPlugin(registry registry.Interface, hugo hugo.Hugo, writer writers.Writer, skipLinkValidation bool) *MarkdownPlugin {
-	return &MarkdownPlugin{
+// New creates a new markdown plugin
+func New(registry registry.Interface, hugo hugo.Hugo, writer writers.Writer, skipLinkValidation bool) *Plugin {
+	return &Plugin{
 		registry:           registry,
 		hugo:               hugo,
 		writer:             writer,
@@ -35,12 +32,12 @@ func NewMarkdownPlugin(registry registry.Interface, hugo hugo.Hugo, writer write
 }
 
 // Name returns the plugin name for identification
-func (p *MarkdownPlugin) Name() string {
+func (p *Plugin) Name() string {
 	return "markdown"
 }
 
 // ManifestTransformations returns transformations to apply during manifest parsing
-func (p *MarkdownPlugin) ManifestTransformations() []manifest.NodeTransformation {
+func (p *Plugin) ManifestTransformations() []manifest.NodeTransformation {
 	return []manifest.NodeTransformation{
 		p.setMarkdownProcessor,
 		p.propagateFrontmatter,
@@ -49,30 +46,45 @@ func (p *MarkdownPlugin) ManifestTransformations() []manifest.NodeTransformation
 }
 
 // FinalNodeStructure is called after manifest resolution with the final document structure
-func (p *MarkdownPlugin) FinalNodeStructure(documentNodes []*manifest.Node) error {
-	// Now we can create the actual markdown processor with the final document structure
-	processor := markdown.NewPlugin(documentNodes, p.registry, p.hugo, p.writer, p.skipLinkValidation)
-	p.markdownProcessor = processor
+func (p *Plugin) FinalNodeStructure(documentNodes []*manifest.Node) error {
+	// Create document worker with the final document structure
+	lr := linkresolver.New(documentNodes, p.registry, p.hugo)
+	p.documentWorker = NewDocumentWorker(lr, p.registry, p.hugo, p.writer, p.skipLinkValidation)
 	return nil
 }
 
 // Processor returns the processor name for node processing
-func (p *MarkdownPlugin) Processor() string {
+func (p *Plugin) Processor() string {
 	return "markdown"
 }
 
 // Process processes a node using the old synchronous method
-func (p *MarkdownPlugin) Process(node *manifest.Node) error {
-	return p.markdownProcessor.Process(node)
+func (p *Plugin) Process(node *manifest.Node) error {
+	// Legacy method - not used since we're using ProcessNew() for channels
+	return nil
 }
 
 // ProcessNew processes a node using the new channel-based method
-func (p *MarkdownPlugin) ProcessNew(node *manifest.Node) []chan nodeplugins.Status {
-	return p.markdownProcessor.ProcessNew(node)
+func (p *Plugin) ProcessNew(node *manifest.Node) []chan plugins.Status {
+	out := make(chan plugins.Status)
+	go func() {
+		defer close(out)
+
+		// Process document using Worker directly - now returns links
+		links, err := p.documentWorker.ProcessNode(context.TODO(), node)
+		if err != nil {
+			out <- plugins.NewStatus(err)
+			return
+		}
+
+		// Send status with collected external links
+		out <- plugins.NewStatusWithLinks(nil, links) // Success
+	}()
+	return []chan plugins.Status{out}
 }
 
 // setMarkdownProcessor sets the processor for markdown files
-func (p *MarkdownPlugin) setMarkdownProcessor(node *manifest.Node, parent *manifest.Node, _ registry.Interface) (bool, error) {
+func (p *Plugin) setMarkdownProcessor(node *manifest.Node, parent *manifest.Node, _ registry.Interface) (bool, error) {
 	if node.Type == "file" && strings.HasSuffix(node.File, ".md") {
 		node.Processor = "markdown"
 	}
@@ -80,7 +92,7 @@ func (p *MarkdownPlugin) setMarkdownProcessor(node *manifest.Node, parent *manif
 }
 
 // propagateFrontmatter propagates frontmatter from parent to child nodes
-func (p *MarkdownPlugin) propagateFrontmatter(node *manifest.Node, parent *manifest.Node, _ registry.Interface) (bool, error) {
+func (p *Plugin) propagateFrontmatter(node *manifest.Node, parent *manifest.Node, _ registry.Interface) (bool, error) {
 	if parent != nil {
 		newFM := map[string]interface{}{}
 		for k, v := range parent.Frontmatter {
@@ -97,7 +109,7 @@ func (p *MarkdownPlugin) propagateFrontmatter(node *manifest.Node, parent *manif
 }
 
 // propagateSkipValidation propagates skip validation settings
-func (p *MarkdownPlugin) propagateSkipValidation(node *manifest.Node, parent *manifest.Node, _ registry.Interface) (bool, error) {
+func (p *Plugin) propagateSkipValidation(node *manifest.Node, parent *manifest.Node, _ registry.Interface) (bool, error) {
 	if parent != nil && parent.Frontmatter != nil {
 		if skipVal := parent.Frontmatter["skip_validation"]; skipVal != nil {
 			if node.Frontmatter == nil {
