+++ pkg/plugins/markdown/tests/docs/images/gardener-docforge-logo.png
