+++ pkg/plugins/markdown/tests/docs/manifest.yaml
