+++ pkg/plugins/persona/persona-filtering.json.tpl
