+++ pkg/plugins/persona/tests/manifests/persona_filtering.yaml
